[p31]Updated functionality based on comments
