--- Empire - The Last Battle/Assets/Unity/Scripts/MonoBehaviours/Inn.cs.meta
@@ -1,12 +1,0 @@
-fileFormatVersion: 2
-guid: 83c1ea112542033439ec2910df70fd5c
-timeCreated: 1450098963
-licenseType: Free
-MonoImporter:
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
